requirements.txt
antiorm==1.2.1
astroid==2.4.2
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.9
click==7.1.2
colorama==0.4.3
cryptography==36.0.1
db==0.1.1
dnspython==2.1.0
email-validator==1.1.3
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-LoginManager==1.1.6
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
idna==3.3
isort==4.3.21
itsdangerous==2.0.1
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.21.4
pandas==1.3.5
pdfminer==20191125
pdfminer.six==20211012
Pillow==8.4.0
pycparser==2.21
pycryptodome==3.12.0
pylint==2.5.3
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.28
toml==0.10.1
urllib3==1.26.7
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==3.0.0
